// repos/sbs-draft-web-main/utils/types/types.ts
export type OwnerProps = {
    availableCredit: number
    availableEthCredit: number
    blueCheckEmail: string
    hasW9: boolean
    isBlueCheckVerified: boolean
    leagues: LeagueProps[]
    numWithdrawals: number
    pendingCredit: number
    withdrawnAmount: {
        2022: number
    }
    pfp: {
        imageUrl: string
        nftContract: string
        displayName: string
    }
}

export type LeagueProps = {
    leagueId: string
    cardId: string
}

export type NFTProps = {
    identifier: string
    collection: string
    contract: string
    token_standard: string
    name: string
    description: string
    image_url: string
    is_disabled: boolean
    is_nsfw: boolean
    metadata_url: string
}

export type LeagueDataProps = {
    available: []
    active: activeLeagueDataProps[]
}

export type LivePlayersProps = {
    id: string
    numPlayers: number
}

export type generatedCardProps = {
    card: activeLeagueDataProps
}

export type activeLeagueDataProps = {
    roster: {
        DST: null | string
        QB: null | string
        RB: null | string
        WR: null | string
        TE: null | string
    }
    _draftType: string
    _cardId: string
    _imageUrl: string
    _level: string
    _ownerId: string
    _leagueId: string
    _leagueDisplayName: string
    _leagueRank: string
    _rank: string
    _weekScore: string
    _seasonScore: string
    numPlayers: number
    prizes: {
        ETH: number
    }
    playoffs: boolean
}
// ...
export type UpdatedRankings = {
    ranking: {
        playerId: string
        rank: number
        score: number
    }[]
}
// ...
export type RankingsProps = {
    playerId: string
    rank: number
    score: number
    stats: {
        averageScore: number
        highestScore: number
        playerId: string
        top5Finishes: number
        adp: number
        byeWeek: string
        playersFromTeam: string[]
    }
}